backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
pillow
python-multipart